requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
reportlab==4.4.4
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
